+++ zabbix_template_sync/__init__.py
@@ -15,4 +15,4 @@ __all__ = [
     "GitLabClient",
     "GitLabAPIError",
 ]
-__version__ = "1.0.3"
+__version__ = "1.0.6"